Consolidate information from definitions in topChunk
This removes `rhs` and `interactionErrors`, which were exclusively used to
represent the output of definitions, and replaces them with a single
`topChunk`.  This simplifies the logic in many places and consolidates the
rendering logic on the parts that chunks use.

We also remove the moribund `Embeditor` mode in this edit, which had bitrotted.
Focus on the text "classic" mode, Chatitor mode, and on the new Examplar mode.