Initial implementation of merge and menu bar for chunks
- Add "merge with previous" button.
- Add menubar over chunks

The focusing stuff is a little hacky -- probably need state of

    currentFocusedChunk : number | undefined

eventually. Right now had to change clicks to mousedowns to capture the click
on the hover menu before blur obscures it and misses the click!

But this is nice because it is a pure CSS-and-local-state solution that relies
on the browser/CM only letting us focus on textarea at a time.

There is a bug with runSegmentsAsync now, which will not clear the run cache
correctly if the *last two* chunks are merged. It will think only the last
chunk has changed and not do a clean re-run that clears out old definitions.

Need to enhance that logic for this case.