Compute, rather than save-and-update, currentFileContents
Rather than trying to always keep currentFileContents up to date in state.ts,
we keep the textual content of the editor only in two places:

- The collection of editor (codemirror) instances for definitions and chunks
- The saved file on disk/localstorage

We compute chunks from string, and current file contents from state, to go back
and forth between the two.

We make it so edits to defintions and all the chunk-manipulating actions
enqueue a save file effect to keep things up to date on disk, always computed
from the current set of editors.

We also standardize a format choice for converting between chat mode an
definitions+chat mode – definitions is the text before the first CHUNKSEP. If
there is no text before the first CHUNKSEP, then it doesn't appear as a chat.

In the future there are lots of design decisions we could make here. I think my
favorite would be to store two files for each program:

- program.arr
- program.parrley (hah hah)

The former has definitions (which could be empty/a trivial string to signal
that there are only chats), and the latter has the interactions/chats.

Developers might choose to e.g. put *.parrley in the .gitignore to distinguish
between their session and the “program”. Each dev can have their own
explorations/chat session. One could even imagine

- program.parrley.1
- program.parrley.2
- ...

For multiple sessions with a UI for picking between them.

This is also relevant for sharing teachpacks, where it's likely that the intent
is *not* to share whatever is in the interactions when a file is imported.
Indeed, this is one of the Grus comments about not liking notebooks – it's
hard to import them because they aren't quite programs. This split would keep
the Repartee benefits (serialize the interactions as a real artifact, etc)
while still clearly defininig what the program is for importing, a set of
definitions, a standalone artifact, etc.