Update editor on edits
I'm not sure if this is a hack or principled.

Because of how React and Uncontrolled CodeMirrors work, when we switch files in
Text/Examplar mode, the same CodeMirror instance is re-used.

We could try to destroy and re-create a CodeMirror here, and that might be the
right thing to do, but I'm not sure how to do that in the right React-y way.

But the main thing is that the `editorDidMount` doesn't trigger when we reset
the value upon changing the CodeMirror, and that was our callback for getting
handles on the underlying CM instances. We pass in "UninitializedEditors" in a
bunch of places, which are just inert objects with a getValue() method that
return a constant. They are intended to always be replaced by live CodeMirrors
as soon as they mount so that they correctly keep fetching up to date contents.

This fix resets definitionsEditor on each change (which will happen right when
the file changes). There will also be lots of spurious updates to state as a
result, but I think that's fine. I don't know if there's a better
“componentDidReceiveProps” or similar to update on, it's not listed in the
docs for react-codemirror2:

https://www.npmjs.com/package/react-codemirror2

And this works pretty well.