The condition to check s-module-global was incorrect
In a program like

include cpo

lists.distinct

The `lists` identifier was not an s-module-global, so `lists` was not
`require`d into the generated code.

In these cases, we should *always* add the URI for the referred module to the
set of modules to directly `require`. A later set operation adds imports for
the appropriate modules with the given names (and doesn't add multiple imports
for the same URI)