Make creation of webworker async, and update dependencies
There was a race because the worker was created before configuring the BFS
filesystem, so sometimes it could start up and not find the FS that was
initialized page-side.

This makes the worker creation happen only after the page-side BFS is fully
initialized (in its callback), and propagates through the async to all the
relevant places that depend on that initialization (many were async anyway).