equalRec should do another loplevel call to equalCore that gets a fresh worklist and wraps up, not get involved in the outer worklist
ls1 = [list-set: [list: "a"], [list: "b"]]
ls2 = [list-set: [list: "b"], [list: "a"]]
ls1 == ls2

would return `false`.

This is because `equalHelp` always returns Equal back to the `list-set._equals`
method, because that's the return value  for structures like lists to signal
“keep going in the worklist”. That's wrong as the recursive equals, because
it gave list-set the wrong anser, **but then the contents of lists were left in
the worklist**, so a later "a" == "b" comparison failed and was combined with
AND as if part of the toplevel structure comparison.