Initialization of checker via a new runtime callback
`initializeCheckContext` will intialize, just for the main module, a check context

We found an issue in CPO/IDE contexts here where we need something like

function mainComplete(uri) {
  if(uri === currentMainURI) { currentMainURI = false; }
}

to *reset* this. Right now in parley/repartee, the main is always set to
`/tmp/include-cpo.arr` because that's what we call to get all the modules
cached.

There's some extant bug with the checker filter that's running *all* checks
anyway, and not respecting this filter (but the mainURL field is also wrong).