Make the ide Failure type be exactly ErrorDisplay from the runtime
There is a little lie here in that all the methods aren't present on compile
errors (which get serialized over a postMessage so they can't). But nothing in
the ide relies on the methods, and typically we render these with functions
anyway, so it's minor. The resulting setup gives very good type-checking in the
IDE and syncs any error representation changes!