Add start/end/editorAtLastRun to ChunkResults
Added a new idea to ChunkResults – they store timestamps for when they
started, when they ended, and what the text was when they ran (as a CodeMirror
doc)

This ensures that code embeds can always find the code they need to (even if
they re-render or the editor changes).

As a nice bonus, in the future this can be made into a widget that shows how
long a chunk took to run.