Make sure to consult cache for built-in modules
OK so I think this bug was introduced a mere 8 years ago:

https://github.com/brownplt/pyret-lang/commit/d3c1652bc9d31c621dddf73b4704a3d7fb1f42f6

I think it was part of changes related to:

https://github.com/brownplt/pyret-lang/commit/e1ecf1dc11c4bc46d609085221917975158aa003

The main thing going on here is we accrued subtle interactions between some
things.

1. builtin modules may have *no associated source*, rendering mtime checks
meaningless. This is most commonly in the npm/server mode, where it would be
silly to ever check an mtime between the compiled version of list.arr.js and
the source of list.arr, controlled by --compiled-read-only-dir
2. builtin modules get copied from some origin location to the `compiled`
directory upon build (this helps make sure that we can build true JS
standalones). This is controlled by --compiled-cache

These ideas were being mucked with at the same time, and I think led to this
refuctoring [sic] that has persisted for some time.

The --compiled-cache should *not* be treated the same way as a
--compiled-read-only-dir (which is what the `link` does).

Note that basedir (the variable in the code that's by default bound to
`"compiled"` or whatever that option happens to be) is used later when doing a
“normal” builtin module location, so that is being propagated and used
correctly in that case.

It is *only* in the case that we know we have a pre-built,
compiled-version-only module that we should skip the mtime infrastructure. (NB:
the pyret-npm server knows to restart itself if it gets updated while it's
running, which handles anoter potential staleness case you might think of based
on that description):

https://github.com/brownplt/pyret-npm/blob/b7a2ed44ddab96951f23a784ccbde52ca3fa9e9d/client-lib.js#L228
).

Or I'm totally wrong and this will cause some other cache issue down the line.
But in my local testing this is a real improvement in some of the cases we've
seen recently where we are working on something in trove/ and it frustratingly
won't have its changes picked up without deleting stuff in compiled/