[experiment] Use record-concat to allow partial options
This is a little weird in that I had to *remove* the annotation checks, because
we can't expression Partial<CSVOptions>.

This means you get less checking if you typo a row (like { hedder-rho: false}).

However, you don't have to type all the options, or use default-options.{ blah
} to change 1 or 2. Also also if you write object literals it's future-proof to
new default options.