WIP: attempt to fix resizing bug when resizing the dialog
For some reason, the <div><svg .../></div> structure for *only Vega renderings* seems to produce a 4.5px height discrepancy,
which doesn't appear either for GCharts renderings or for this same markup when used in a separate html file.  I can't track
the root cause down, so I've worked around it by only resizing the Vega chart when the dialog fires its resizeStop event.

The resizing still isn't perfect, but it's better than it was.