New npm() import type
A new builtin module (for internal use most likely) called require-util exposes
Node's require function. Note that this won't quite work in non-Node contexts
without a working require. It can be polyfilled with
https://github.com/browserify/resolve and ZenFS. However, ZenFS atop VScode
FileSystem isn't a thing yet, so it's hard to implement it fully in the
extension without some more work.

Interestingly this may be a place where splitting off the behavior for the
desktop extension and the web extension makes sense. The desktop extension
should more easily have access to the right kind of require, and is the place
where someone could plausibly run `npm install` – in the web context there's
not even a way to meaningfully run npm. This would give the desktop extension
(and codespaces!) the ability to use `npm`/`node_modules` infrastructure to
package and distribute Pyret files.

For now, this is a good CLI solution