Change npm imports to use browserify's resolve()
This uses https://github.com/browserify/resolve for resolution. A few reasons:

1. It has a clean API for giving a `basedir` argument
2. It is pure JavaScript and (given a working `fs` and `path` implementation),
can run in a webview/browser tab, rather than being node-internal

Then, use the basedir configuration to make sure `npm` imports are resolved
relative to the module containing the import (not relative to the compiler).

Also update the test to have a handwritten node_modules with pyret-test-files.
This ensures that when we run the _compiler_, whose working directory is the
root of the repo, it must be using this new infrastructure to find the
pyret-test-files module.