Update Image polyfill.
OK, this is super weird. I'm not sure how this was working before, but it
certainly was, unless somehow node-canvas behaves differently in webviews than
in real browser tabs. Also I'm not sure how new-image-things ever passed on
Travis.

But it sure seems like node-canvas does not in fact provide an Image
constructor when run in-browser -- canvas.Image is clearly undefined and only
the default browser canvas APIs are there when we import it this way.

So, we have to do the polyfill bit ourselves and make sure `Image` is correctly
set up as the nonstandard canvas.Image from node-canvas when offline and to the
browser's window.Image in browsers