Make this backwards-compatible with clients that didn't know about orig-headers
To avoid potential clashes with different clients, it's better if the new spec
for these tuples is that the new field is at the end, and to allow the old or
new length.

Also pragmatically this avoids coming up with a good answer for gdrive-sheets
while still enjoying this runtime change for CSV loading.